Moving rendering to tripController
