+++ src/main.js
@@ -1,63 +1,12 @@
-import Info from '../src/components/info';
 import TotalCost from '../src/components/total-cost';
 import Menu from '../src/components/menu';
 import Filters from '../src/components/filters';
-import TripSorting from '../src/components/sortings';
-import Card from '../src/components/card';
-import CardEditing from '../src/components/card-editing';
-import NoCards from '../src/components/no-cards';
-import Days from '../src/components/days';
-import Day from '../src/components/day';
-import {render, RenderPosition, replace} from './utils/render';
+import TripController from '../src/controllers/trip-controller';
+import {render, RenderPosition} from './utils/render';
 
-import filters from "./mock/filter";
-import menuItems from "./mock/menu";
-import generateCard from "./mock/card";
-
-const CARDS_COUNT = 3;
-const points = new Array(CARDS_COUNT).fill(``).map((point) => {
-  point = generateCard();
-  return point;
-});
-
-const groupPointsByDate = (items) => {
-  const pointDates = Array.from(new Set(items.map((point) => {
-    return point.date;
-  })));
-
-  return pointDates.map((pointDate, i) => {
-    return {
-      day: {
-        number: i + 1,
-        date: pointDate,
-      },
-      points: items.filter((point) => point.date === pointDate)
-    };
-  });
-};
-
-const renderPoint = (point, dayElement) => {
-  const onEscKeyDown = (evt) => {
-    const isEscKey = evt.key === `Escape` || evt.key === `Esc`;
-
-    if (isEscKey) {
-      replace(pointComponent, pointEditComponent);
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    }
-  };
-
-  const pointComponent = new Card(point);
-  const pointEditComponent = new CardEditing(point);
-
-  pointComponent.setRollupButtonClickHandler(() => {
-    replace(pointEditComponent, pointComponent);
-    document.addEventListener(`keydown`, onEscKeyDown);
-  });
-
-  pointEditComponent.setSubmitHandler(() => replace(pointComponent, pointEditComponent));
-
-  render(dayElement, pointComponent.getElement(), RenderPosition.BEFOREEND);
-};
+import filters from './mock/filter';
+import menuItems from './mock/menu';
+import points from './mock/points';
 
 const infoElement = document.querySelector(`.trip-info`);
 const controlsElement = document.querySelector(`.trip-controls`);
@@ -67,19 +16,6 @@ render(infoElement, new TotalCost(points).getElement(), RenderPosition.BEFOREEND
 render(controlsElement, new Menu(menuItems).getElement(), RenderPosition.BEFOREEND);
 render(controlsElement, new Filters(filters).getElement(), RenderPosition.BEFOREEND);
 
-if (points.length) {
-  render(infoElement, new Info(points).getElement(), RenderPosition.AFTERBEGIN);
-  render(eventsElement, new TripSorting().getElement(), RenderPosition.BEFOREEND);
-  render(eventsElement, new Days().getElement(), RenderPosition.BEFOREEND);
+const tripController = new TripController(eventsElement, points);
 
-  const daysElement = eventsElement.querySelector(`.trip-days`);
-  const groupedByDatePoints = groupPointsByDate(points);
-
-  groupedByDatePoints.forEach((date) => {
-    render(daysElement, new Day(date.day).getElement(), RenderPosition.BEFOREEND);
-    const dayElement = eventsElement.querySelector(`.trip-days__item:nth-child(${date.day.number}) .trip-events__list`);
-    date.points.forEach((point) => renderPoint(point, dayElement));
-  });
-} else {
-  render(eventsElement, new NoCards().getElement(), RenderPosition.BEFOREEND);
-}
+tripController.render();
